Version-3 May 3, 2024 08:55 pm

+++ lib/logic/leaderboard/leaderboard_ql.dart
@@ -54,8 +54,11 @@ class LeaderBoardQL {
         conflictAlgorithm: ConflictAlgorithm.replace,
       );
 
-  Future delete() async =>
+  Future delete() async {
+    if (await databaseFactory.databaseExists(await getDatabasesPath())) {
       deleteDatabase(join(await getDatabasesPath(), '$_tableName.db'));
+    }
+  }
 }
 
 /* Future<List<Player>> getAllPlayersSortedByRank() async {
